requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
openai==1.3.0
beautifulsoup4==4.12.2
